+++ src/components/app/app.tsx
@@ -7,19 +7,22 @@ import LoginPage from '../../pages/login-page/login-page';
 import NotFoundPage from '../../pages/not-found-page/not-found-page';
 import OfferPage from '../../pages/offer-page/offer-page';
 import PrivateRoute from '../private-route/private-route';
+import FavoritePage from '../../pages/favorites-page/favorites-page';
+
+import { OfferType } from '../../types/offer';
 
 type AppProps = {
-  placesCount: number;
+  offers: OfferType[];
 }
 
-function App(props: AppProps): JSX.Element {
+function App({offers}: AppProps): JSX.Element {
   return (
     <HelmetProvider>
       <BrowserRouter>
         <Routes>
           <Route
             path={AppRoute.Main}
-            element={<MainPage {...props}/>}
+            element={<MainPage offers={offers} />}
           />
           <Route
             path={AppRoute.Login}
@@ -31,13 +34,13 @@ function App(props: AppProps): JSX.Element {
               <PrivateRoute
                 authorizationStatus={AuthorizationStatus.Auth}
               >
-                <MainPage {...props}/>
+                <FavoritePage offers={offers} />
               </PrivateRoute>
             }
           />
           <Route
-            path={AppRoute.Offer}
-            element={<OfferPage />}
+            path={`${AppRoute.Offer}/:offerId`}
+            element={<OfferPage offers={offers} />}
           />
           <Route
             path="*"
