+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from '../src/components/app/app';
-import {PLACES_COUNT} from './common/const';
+import {offers} from './mock/offers';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -9,6 +9,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App placesCount={PLACES_COUNT}/>
+    <App offers = {offers}/>
   </React.StrictMode>
 );
