+++ src/types/offer-city.ts
@@ -1,0 +1,7 @@
+import { OfferLocation } from '../types/offer-location';
+import { CityName } from '../common/const';
+
+export type OfferCity = {
+  location: OfferLocation;
+  name: CityName;
+};
