+++ src/types/offer-location.ts
@@ -1,0 +1,5 @@
+export type OfferLocation = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
